Add visit_class and refactor redundant traversals when both adjustFileObjs and adjustDirObjs are called on the same objects with the same function.
