+++ tests/test_pack.py
@@ -7,7 +7,7 @@ import cwltool.pack
 import cwltool.workflow
 from cwltool.load_tool import fetch_document, validate_document
 from cwltool.main import makeRelative
-from cwltool.process import adjustFileObjs, adjustDirObjs
+from cwltool.pathmapper import adjustFileObjs, adjustDirObjs
 from .util import get_data
 
 class TestPack(unittest.TestCase):
